Fixed code after rivew
Fixed:
styles
sprite for close button
refactored modal.js
added modalExercise.ts
refactored addToFavorite.js
