+++ src/css/components/modal-animation.css
@@ -6,6 +6,14 @@
 .modal[open] {
   opacity: 1;
   transform: translateY(0px);
+
+  /*   Before-open state  */
+  /* Needs to be after the previous dialog[open] rule to take effect,
+    as the specificity is the same */
+  @starting-style {
+    opacity: 0;
+    transform: translateY(10%);
+  }
 }
 
 /*   Closed state of the dialog   */
@@ -18,16 +26,6 @@
     display var(--transition-duration) var(--transition-function) allow-discrete;
 }
 
-/*   Before-open state  */
-/* Needs to be after the previous dialog[open] rule to take effect,
-    as the specificity is the same */
-@starting-style {
-  .modal[open] {
-    opacity: 0;
-    transform: translateY(10%);
-  }
-}
-
 /* Transition the :backdrop when the dialog modal is promoted to the top layer */
 .modal::backdrop {
   background-color: rgb(0 0 0 / 0%);
@@ -38,13 +36,11 @@
 
 .modal[open]::backdrop {
   background-color: rgb(0 0 0 / 25%);
+
+  @starting-style {
+    background-color: rgb(0 0 0 / 0%);
+  }
 }
 
 /* This starting-style rule cannot be nested inside the above selector
 because the nesting selector cannot represent pseudo-elements. */
-
-@starting-style {
-  .modal[open]::backdrop {
-    background-color: rgb(0 0 0 / 0%);
-  }
-}
