+++ src/css/components/modal.css
@@ -1,3 +1,7 @@
+body {
+  background-color: rgb(36, 36, 36);
+}
+
 .modal {
   border: none;
   padding: 0;
@@ -46,12 +50,6 @@
     height: 100%;
     min-width: 270px;
   }
-}
-
-img {
-  display: block;
-  width: 100%;
-  height: auto;
 }
 
 .modal-close-button {
