+++ src/index.html
@@ -30,6 +30,6 @@
     <load src="./partials/modal.html" />
 
     <script type="module" src="./ts/pages/home.ts"></script>
-    <script type="module" src="./ts/components/modal/modal.ts"></script>
+    <script type="module" src="./ts/components/modal/modalExercise.ts"></script>
   </body>
 </html>
