+++ src/partials/modal.html
@@ -1,20 +1,8 @@
-<dialog class="modal">
+<dialog class="modal hidden" hidden>
   <div class="modal-container">
     <button class="modal-close-button" type="button" data-close-element>
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-      >
-        <path
-          d="M17 7L7 17M7 7L17 17"
-          stroke="#F4F4F4"
-          stroke-width="1.5"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        />
+      <svg width="24" height="24" class="icon stroke-icon">
+        <use href="/images/sprite.svg#icon-x" data-favorite-icon />
       </svg>
     </button>
     <div class="modal-card">
@@ -46,7 +34,7 @@
           <!-- #region star fill -->
           <svg class="hidden-svg-clip">
             <defs>
-              <clipPath id="cut-off-star">
+              <clipPath data-cut-off-star>
                 <rect x="0" y="0" width="7" height="13" data-cut-percent />
               </clipPath>
             </defs>
